modified:   3rdPartLib/DBow3/src/DescManip.h 	modified:   launch/vo.launch 	modified:   src/feature_dem.cpp 	modified:   src/include/common.h 	modified:   src/include/feature_dem.h 	modified:   src/vo_loopclosing.cpp 	modified:   src/vo_tracking.cpp 	deleted:    voc/ORBvoc.bin 	voc/

+++ 3rdPartLib/DBow3/src/DescManip.h
@@ -81,7 +81,7 @@ uint32_t DescManip::distance_8uc1(const cv::Mat &a, const cv::Mat &b){
          pb = b.ptr<uint64_t>();
 
          uint64_t v, ret = 0;
-         int n=a.cols / sizeof(uint64_t);
+         uint64_t n=a.cols / sizeof(uint64_t);
          for(size_t i = 0; i < n; ++i, ++pa, ++pb)
          {
            v = *pa ^ *pb;
